refactor: update type imports
